[WIP] Eliminate gyro yaw drift by keeping spine trackers aligned with the headset
Most of the time, the player's body faces in the same direction as their head. We tend to stand facking forward, walk forward, even sit facing forward. We may look to the side or twist our bodies, but we usually return to the forward position. Therefore, we expect the hip and chest trackers to be aligned with the headset.

We constantly check whether the hip and chest trackers have the same yaw as the headset. If there is a difference, we add a tiny yaw rotation to reduce the difference. The tiny rotation is not noticeable, unless the user looks to the side for an extended time.

A typical calibrated BMI270 tracker can have a gryo at-rest-bias of 0.1deg/s (0.002rad/s) due to temperature changes over play time. In 10min, you may have rotated up to 60deg from your actual rotation. This is very difficult to calibrate away even with temperature calibration because of hysteresis. VQF helps by damping the bias, but doesn't have enough information to completely eliminate the bias.

After testing this in VRC for 30mins, my chest and hip trackers were still perfectly aligned, even though there was an actual gyro yaw drift of 15deg.