[Spine Yaw Compensation][v3] Better alignment of trackers along Y-Z axes
After a reset, all the trackers are aligned along their Y-Z axes. Previously, we were trying to align the yaw, which only works when the skeleton is vertical. In v3, we try to keep all the trackers aligned along their Y-Z axes, by adding small yaw compensations. This works in non-vertical orientations such as sitting down and lying down.

1. We try to align the upper body trackers to the head, as long as the head isn't looking off to the side.
2. We try to align the legs to the upper body, as long as the leg isn't twisted off to the side.
3. We try to always align the leg trackers to each other, since they should be mostly locked into the same Y-Z plane by the knee.
4. We try to always align the feet trackers to the leg, because the feet often get completely out of alignment when going from standing to lying down.